Getting list of recently used categories
This is actually kinda limited because it looks at the entry date and
not the date that the entry was actually added. Further, there's only
date precision on entries but recently used could benefit from more
precision.

Should consider adding a "datetime added/edited" column that could be
used for this function. This would then allow us to have the recently
used category list working as expected.